backend/requirements.txt
black==24.2.0
Django==5.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
supabase==2.7.4
requests==2.32.3
django-cors-headers==4.1
pillow==10.4.0
tensorflow==2.17.0
matplotlib==3.9.2
numpy==1.26.4
opencv-python==4.10.0.84
langchain-google-genai==2.0.0
langchain==0.3.1
langchain-community==0.3.1
pypdf==5.0.1
chromadb==0.5.11
langchain-chroma==0.1.4
